requirements.txt
playwright>=1.50.0
requests>=2.32.3
schedule>=1.2.2
python-dotenv>=1.0.1
colorlog>=6.8.2
pytest>=8.2.2
pytest-asyncio>=0.23.7
flask>=3.1
openpyxl>=3.1
reportlab>=4.0

# ML-based PDF extraction
transformers>=4.35.0
torch>=2.0.0
pdfplumber>=0.10.0
Pillow>=10.0.0
paddleocr>=2.7.0.0

# Data handling
pandas>=2.0.0
PyPDF2>=4.0.0
numpy>=1.24.0
